on scrren keys resizing
